misc/HtmlTestSuite/results/test-0499.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><meta charset="utf-8"/>
<title>CSS Grid Layout Test: Minimum size of grid items</title>
<link href="mailto:rego@igalia.com" rel="author" title="Manuel Rego Casasnovas"/>
<link href="http://www.w3.org/TR/css-grid-1/#min-size-auto" rel="help" title="6.6. Implied Minimum Size of Grid Items"/>
<link href="http://www.w3.org/TR/css-grid-1/#algo-track-sizing" rel="help" title="12.3. Track Sizing Algorithm"/>
<meta content="Checks how automatic minimum size of images affect to the calculation of the grid container size and the grid tracks. Verifies the sizing of the image in the different cases too." name="assert"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<link href="/fonts/ahem.css" rel="stylesheet" type="text/css"/>
<style>
  .grid {
    display: inline-grid;
    grid-template-rows: auto;
    grid-template-columns: auto;
    font: 25px/1 Ahem;
  }

  .grid2rows {
    grid-template-rows: auto auto;
  }

  .constrainedGrid {
    width: 10px;
    height: 10px;
  }

  .justifyContentStart {
    justify-content: start;
  }

  .width200px {
    width: 200px;
  }

  .width100percent {
    width: 100%;
  }
</style>

<script>
  function checkGridSizeTracksAndImageSize(gridId, imgId, gridWidth, gridHeight, gridColumns, gridRows, imgWidth, imgHeight) {
    const gridStyle = getComputedStyle(document.getElementById(gridId));
    const imgStyle = getComputedStyle(document.getElementById(imgId));

    test(
      function() {
        assert_equals(gridStyle.width, gridWidth);
      }, gridId + ".width");
    test(
      function() {
        assert_equals(gridStyle.height, gridHeight);
      }, gridId + ".height");
    test(
      function() {
        assert_equals(gridStyle.gridTemplateColumns, gridColumns);
      }, gridId + ".gridTemplateColumns");
    test(
      function() {
        assert_equals(gridStyle.gridTemplateRows, gridRows);
      }, gridId + ".gridTemplateRows");
    test(
      function() {
        assert_equals(imgStyle.width, imgWidth);
      }, imgId + ".width");
    test(
      function() {
        assert_equals(imgStyle.height, imgHeight);
      }, imgId + ".height");
  }

  setup({ explicit_done: true });

  function runTests() {
    checkGridSizeTracksAndImageSize("grid-1", "img-1", "200px", "200px", "200px", "200px", "200px", "200px");
    checkGridSizeTracksAndImageSize("grid-2", "img-2", "10px", "10px", "200px", "200px", "200px", "200px");
    checkGridSizeTracksAndImageSize("grid-3", "img-3", "200px", "200px", "200px", "200px", "200px", "200px");
    checkGridSizeTracksAndImageSize("grid-4", "img-4", "200px", "10px", "200px", "200px", "200px", "200px");
    checkGridSizeTracksAndImageSize("grid-5", "img-5", "200px", "50px", "50px", "50px", "50px", "50px");
    checkGridSizeTracksAndImageSize("grid-6", "img-6", "200px", "10px", "50px", "50px", "50px", "50px");
    checkGridSizeTracksAndImageSize("grid-7", "img-7", "200px", "225px", "200px", "200px 25px", "200px", "200px");
    checkGridSizeTracksAndImageSize("grid-8", "img-8", "10px", "10px", "200px", "200px 25px", "200px", "200px");
    checkGridSizeTracksAndImageSize("grid-9", "img-9", "200px", "225px", "200px", "200px 25px", "200px", "200px");
    checkGridSizeTracksAndImageSize("grid-10", "img-10", "200px", "10px", "200px", "200px 25px", "200px", "200px");
    checkGridSizeTracksAndImageSize("grid-11", "img-11", "200px", "125px", "100px", "100px 25px", "100px", "100px");
    checkGridSizeTracksAndImageSize("grid-12", "img-12", "200px", "10px", "100px", "100px 25px", "100px", "100px");
    checkGridSizeTracksAndImageSize("grid-13", "img-13", "200px", "200px", "200px", "200px", "200px", "200px");
    checkGridSizeTracksAndImageSize("grid-14", "img-14", "10px", "10px", "200px", "200px", "200px", "200px");
    checkGridSizeTracksAndImageSize("grid-15", "img-15", "200px", "200px", "200px", "200px", "200px", "200px");
    checkGridSizeTracksAndImageSize("grid-16", "img-16", "200px", "10px", "200px", "200px", "200px", "200px");
    checkGridSizeTracksAndImageSize("grid-17", "img-17", "200px", "200px", "200px", "200px", "200px", "200px");
    checkGridSizeTracksAndImageSize("grid-18", "img-18", "200px", "10px", "200px", "200px", "200px", "200px");
    checkGridSizeTracksAndImageSize("grid-19", "img-19", "200px", "225px", "200px", "200px 25px", "200px", "200px");
    checkGridSizeTracksAndImageSize("grid-20", "img-20", "10px", "10px", "200px", "200px 25px", "200px", "200px");
    checkGridSizeTracksAndImageSize("grid-21", "img-21", "200px", "225px", "200px", "200px 25px", "200px", "200px");
    checkGridSizeTracksAndImageSize("grid-22", "img-22", "200px", "10px", "200px", "200px 25px", "200px", "200px");
    checkGridSizeTracksAndImageSize("grid-23", "img-23", "200px", "225px", "200px", "200px 25px", "200px", "200px");
    checkGridSizeTracksAndImageSize("grid-24", "img-24", "200px", "10px", "200px", "200px 25px", "200px", "200px");

    done();
  }
</script>

</head><body onload="document.fonts.ready.then(() =&gt; { runTests(); })">
<div id="log"/>

<!-- Grids with only a 50x50 image as grid item. -->

<div class="grid" id="grid-1">
  <img class="width200px" id="img-1" src="support/50x50-green.png"/>
</div>

<div class="grid constrainedGrid" id="grid-2">
  <img class="width200px" id="img-2" src="support/50x50-green.png"/>
</div>

<div class="grid width200px" id="grid-3">
  <img class="width100percent" id="img-3" src="support/50x50-green.png"/>
</div>

<div class="grid width200px constrainedGrid" id="grid-4">
  <img class="width100percent" id="img-4" src="support/50x50-green.png"/>
</div>

<div class="grid width200px justifyContentStart" id="grid-5">
  <img class="width100percent" id="img-5" src="support/50x50-green.png"/>
</div>

<div class="grid width200px constrainedGrid justifyContentStart" id="grid-6">
  <img class="width100percent" id="img-6" src="support/50x50-green.png"/>
</div>

<!-- Grids with a 50x50 image as grid item and a 100x25 text grid item. -->

<div class="grid grid2rows" id="grid-7">
  <img class="width200px" id="img-7" src="support/50x50-green.png"/>
  <div>ITEM</div>
</div>

<div class="grid grid2rows constrainedGrid" id="grid-8">
  <img class="width200px" id="img-8" src="support/50x50-green.png"/>
  <div>ITEM</div>
</div>

<div class="grid grid2rows width200px" id="grid-9">
  <img class="width100percent" id="img-9" src="support/50x50-green.png"/>
  <div>ITEM</div>
</div>

<div class="grid grid2rows width200px constrainedGrid" id="grid-10">
  <img class="width100percent" id="img-10" src="support/50x50-green.png"/>
  <div>ITEM</div>
</div>

<div class="grid grid2rows width200px justifyContentStart" id="grid-11">
  <img class="width100percent" id="img-11" src="support/50x50-green.png"/>
  <div>ITEM</div>
</div>

<div class="grid grid2rows width200px constrainedGrid justifyContentStart" id="grid-12">
  <img class="width100percent" id="img-12" src="support/50x50-green.png"/>
  <div>ITEM</div>
</div>

<!-- Grids with only a 500x500 image as grid item. -->

<div class="grid" id="grid-13">
  <img class="width200px" id="img-13" src="support/500x500-green.png"/>
</div>

<div class="grid constrainedGrid" id="grid-14">
  <img class="width200px" id="img-14" src="support/500x500-green.png"/>
</div>

<div class="grid width200px" id="grid-15">
  <img class="width100percent" id="img-15" src="support/500x500-green.png"/>
</div>

<div class="grid width200px constrainedGrid" id="grid-16">
  <img class="width100percent" id="img-16" src="support/500x500-green.png"/>
</div>

<div class="grid width200px justifyContentStart" id="grid-17">
  <img class="width100percent" id="img-17" src="support/500x500-green.png"/>
</div>

<div class="grid width200px constrainedGrid justifyContentStart" id="grid-18">
  <img class="width100percent" id="img-18" src="support/500x500-green.png"/>
</div>

<!-- Grids with a 500x500 image as grid item and a 100x25 text grid item. -->

<div class="grid grid2rows" id="grid-19">
  <img class="width200px" id="img-19" src="support/500x500-green.png"/>
  <div>ITEM</div>
</div>

<div class="grid grid2rows constrainedGrid" id="grid-20">
  <img class="width200px" id="img-20" src="support/500x500-green.png"/>
  <div>ITEM</div>
</div>

<div class="grid grid2rows width200px" id="grid-21">
  <img class="width100percent" id="img-21" src="support/500x500-green.png"/>
  <div>ITEM</div>
</div>

<div class="grid grid2rows width200px constrainedGrid" id="grid-22">
  <img class="width100percent" id="img-22" src="support/500x500-green.png"/>
  <div>ITEM</div>
</div>

<div class="grid grid2rows width200px justifyContentStart" id="grid-23">
  <img class="width100percent" id="img-23" src="support/500x500-green.png"/>
  <div>ITEM</div>
</div>

<div class="grid grid2rows width200px constrainedGrid justifyContentStart" id="grid-24">
  <img class="width100percent" id="img-24" src="support/500x500-green.png"/>
  <div>ITEM</div>
</div>
</body>
</html>
